server/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pydantic
pillow